Learning/158_AudioTranscript/requirements.txt
pydub
SpeechRecognition
openai-whisper
vosk